remove switch from route.js.jsx because it tries to match path
