+++ client/global/components/routing/LoginRoute.js.jsx
@@ -11,21 +11,15 @@ import { Redirect, Route } from 'react-router-dom';
 import Auth from '../../utils/auth';
 
 export default class LoginRoute extends React.Component {
-
   constructor(props) {
     super(props);
   }
 
   render() {
     if(Auth.notLoggedIn()) {
-      return <Redirect to={{
-          pathname: "/user/login"
-          , state: { from: this.props.location }
-        }}
-      />
+      return <Redirect to={{pathname: "/user/login", state: { from: this.props.location }}}/>
     } else {
       return <Route {...this.props} />
     }
   }
-
 }
