+++ client/modules/moduleRoutes.js
@@ -4,10 +4,5 @@
  * NOTE: this facilitates adding routes via the CLI
  */
 
-
-// export { default as admin } from './admin/adminRoutes.js.jsx';
-// export { default as product } from './product/productPath.js.jsx';
-// export { default as user } from './user/userRoutes.js.jsx';
-// export { default as userProfile } from './user/userProfileRoutes.js.jsx';
-
 export { default as ProductRouter } from './product/ProductRouter.js.jsx';
+export { default as UserRouter } from './user/UserRouter.js.jsx';
